--- Apollo Deliver/lib/Dto/JwtResponse.dart
@@ -1,23 +1,0 @@
-class JwtResponse {
-  final String accessToken;
-  final String refreshToken;
-
-  JwtResponse({
-    required this.accessToken,
-    required this.refreshToken,
-  });
-
-  factory JwtResponse.fromJson(Map<String, dynamic> json) {
-    return JwtResponse(
-      accessToken: json['accessToken'],
-      refreshToken: json['refreshToken'],
-    );
-  }
-
-  Map<String, dynamic> toJson() {
-    return {
-      'accessToken': accessToken,
-      'refreshToken': refreshToken,
-    };
-  }
-}
