+++ Apollo Deliver/lib/Screens/MakeColor/OnboardingInfo.dart
@@ -1,0 +1,7 @@
+class OnboardingInfo{
+  final String title;
+  final String descriptions;
+  final String image;
+
+  OnboardingInfo({required this.title, required this.descriptions, required this.image});
+}
